+++ deltafi-ui/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Errors can now be acknowledged
 - Added Show/Hide Acknowledged button to Errors page
 - Date range selected has been removed from Errors page
+- DeltaFile viewer displays domains and enrichment data
 
 ## [0.16.3] - 2022-01-19
 ### Added